+++ tests/testthat/test-panelr.R
@@ -1,4 +1,6 @@
-if (require("testthat") && require("insight") && require("panelr")) {
+if (require("testthat") &&
+  require("insight") &&
+  require("panelr")) {
   context("insight, panelr")
 
   data("WageData")
@@ -12,11 +14,29 @@ if (require("testthat") && require("insight") && require("panelr")) {
   })
 
   test_that("find_predictors", {
-    expect_identical(find_predictors(m1), list(conditional = c("union", "wks"), instruments = c("blk", "fem")))
-    expect_identical(find_predictors(m1, flatten = TRUE), c("union", "wks", "blk", "fem"))
-    expect_identical(find_predictors(m1, effects = "random"), list(random = "lag(union)"))
+    expect_identical(
+      find_predictors(m1),
+      list(
+        conditional = c("union", "wks"),
+        instruments = c("blk", "fem")
+      )
+    )
+    expect_identical(
+      find_predictors(m1, flatten = TRUE),
+      c("union", "wks", "blk", "fem")
+    )
+    expect_identical(
+      find_predictors(m1, effects = "random"),
+      list(random = "lag(union)")
+    )
 
-    expect_identical(find_predictors(m2), list(conditional = c("union", "wks"), instruments = c("blk", "fem")))
+    expect_identical(
+      find_predictors(m2),
+      list(
+        conditional = c("union", "wks"),
+        instruments = c("blk", "fem")
+      )
+    )
     expect_identical(find_predictors(m2, effects = "random"), list(random = "id"))
   })
 
@@ -39,8 +59,14 @@ if (require("testthat") && require("insight") && require("panelr")) {
   })
 
   test_that("get_predictors", {
-    expect_equal(colnames(get_predictors(m1)), c("lag(union)", "wks", "blk", "fem"))
-    expect_equal(colnames(get_predictors(m2)), c("lag(union)", "wks", "blk", "fem"))
+    expect_equal(
+      colnames(get_predictors(m1)),
+      c("lag(union)", "wks", "blk", "fem")
+    )
+    expect_equal(
+      colnames(get_predictors(m2)),
+      c("lag(union)", "wks", "blk", "fem")
+    )
   })
 
   test_that("link_inverse", {
@@ -51,12 +77,34 @@ if (require("testthat") && require("insight") && require("panelr")) {
     expect_equal(nrow(get_data(m1)), 3570)
     expect_equal(
       colnames(get_data(m1)),
-      c("lwage", "id", "t", "lag(union)", "wks", "blk", "fem", "imean(lag(union))",
-        "imean(wks)", "lag(union)*blk", "imean(lag(union):blk)", "lag(union):blk")
+      c(
+        "lwage",
+        "id",
+        "t",
+        "lag(union)",
+        "wks",
+        "blk",
+        "fem",
+        "imean(lag(union))",
+        "imean(wks)",
+        "lag(union)*blk",
+        "imean(lag(union):blk)",
+        "lag(union):blk"
+      )
     )
     expect_equal(
       colnames(get_data(m2)),
-      c("lwage", "id", "t", "lag(union)", "wks", "blk", "fem", "imean(lag(union))", "imean(wks)")
+      c(
+        "lwage",
+        "id",
+        "t",
+        "lag(union)",
+        "wks",
+        "blk",
+        "fem",
+        "imean(lag(union))",
+        "imean(wks)"
+      )
     )
   })
 
@@ -82,11 +130,33 @@ if (require("testthat") && require("insight") && require("panelr")) {
   })
 
   test_that("find_variables", {
-    expect_equal(find_variables(m1), list(response = "lwage", conditional = c("union", "wks"), instruments = c("blk", "fem"), random = "lag(union)"))
-    expect_equal(find_variables(m1, flatten = TRUE), c("lwage", "union", "wks", "blk", "fem", "lag(union)"))
+    expect_equal(
+      find_variables(m1),
+      list(
+        response = "lwage",
+        conditional = c("union", "wks"),
+        instruments = c("blk", "fem"),
+        random = "lag(union)"
+      )
+    )
+    expect_equal(
+      find_variables(m1, flatten = TRUE),
+      c("lwage", "union", "wks", "blk", "fem", "lag(union)")
+    )
 
-    expect_equal(find_variables(m2), list(response = "lwage", conditional = c("union", "wks"), instruments = c("blk", "fem"), random = "id"))
-    expect_equal(find_variables(m2, flatten = TRUE), c("lwage", "union", "wks", "blk", "fem", "id"))
+    expect_equal(
+      find_variables(m2),
+      list(
+        response = "lwage",
+        conditional = c("union", "wks"),
+        instruments = c("blk", "fem"),
+        random = c("t", "id")
+      )
+    )
+    expect_equal(
+      find_variables(m2, flatten = TRUE),
+      c("lwage", "union", "wks", "blk", "fem", "t", "id")
+    )
   })
 
   test_that("n_obs", {
@@ -102,7 +172,16 @@ if (require("testthat") && require("insight") && require("panelr")) {
     expect_equal(
       find_parameters(m1),
       list(
-        conditional = c("(Intercept)", "imean(lag(union))", "imean(wks)", "lag(union)", "wks", "blk", "fem", "lag(union):blk"),
+        conditional = c(
+          "(Intercept)",
+          "imean(lag(union))",
+          "imean(wks)",
+          "lag(union)",
+          "wks",
+          "blk",
+          "fem",
+          "lag(union):blk"
+        ),
         random = list(id = "(Intercept)")
       )
     )
@@ -111,26 +190,79 @@ if (require("testthat") && require("insight") && require("panelr")) {
     expect_equal(
       find_parameters(m2),
       list(
-        conditional = c("(Intercept)", "imean(lag(union))", "imean(wks)", "lag(union)", "wks", "blk", "fem"),
-        random = list(id = c("(Intercept)", "t")))
+        conditional = c(
+          "(Intercept)",
+          "imean(lag(union))",
+          "imean(wks)",
+          "lag(union)",
+          "wks",
+          "blk",
+          "fem"
+        ),
+        random = list(id = c("(Intercept)", "t"))
+      )
     )
-
   })
+
+
+  test_that("get_parameters", {
+    expect_equal(
+      get_parameters(m1),
+      data.frame(
+        parameter = c(
+          "(Intercept)",
+          "imean(lag(union))",
+          "imean(wks)",
+          "lag(union)",
+          "wks",
+          "blk",
+          "fem",
+          "lag(union):blk"
+        ),
+        estimate = c(
+          6.59813245629044,
+          -0.0279959204722801,
+          0.00438047648390025,
+          0.0582474262882615,
+          -0.00163678667081885,
+          -0.229414915661438,
+          -0.441756913071962,
+          -0.127319623945541
+        ),
+        stringsAsFactors = FALSE
+      ),
+      tolerance = 1e-4
+    )
+  })
+
 
   test_that("find_terms", {
     expect_equal(
       find_terms(m1),
-      list(response = "lwage", conditional = c("lag(union)", "wks"),
-           instruments = c("blk", "fem"), random = c("blk", "lag(union)"))
+      list(
+        response = "lwage",
+        conditional = c("lag(union)", "wks"),
+        instruments = c("blk", "fem"),
+        random = c("blk", "lag(union)")
       )
+    )
     expect_equal(
       find_terms(m2),
-      list(response = "lwage", conditional = c("lag(union)", "wks"),
-           instruments = c("blk", "fem"), random = c("blk", "t", "id"))
+      list(
+        response = "lwage",
+        conditional = c("lag(union)", "wks"),
+        instruments = c("blk", "fem"),
+        random = c("blk", "t", "id")
+      )
     )
   })
 
   test_that("is_multivariate", {
     expect_false(is_multivariate(m1))
+  })
+
+  test_that("find_statistic", {
+    expect_identical(find_statistic(m1), "t-statistic")
+    expect_identical(find_statistic(m2), "t-statistic")
   })
 }